modbus: make iface an signed integer
The Modbus API defines the interface as a signed integer.

However, there is an inconsistency in the Modbus API with the function
modbus_disable() that use an unsigned integer.

This updates the modbus_disable() prototype to use a signed integer iface
and applies the change in the direct subsequent call core function
modbus_get_context() and in the indirect functions modbus_init_server() and
modbus_init_client().

+++ include/zephyr/modbus/modbus.h
@@ -559,7 +559,7 @@ int modbus_init_client(const int iface, struct modbus_iface_param param);
  *
  * @retval           0 If the function was successful
  */
-int modbus_disable(const uint8_t iface);
+int modbus_disable(const int iface);
 
 /**
  * @brief Submit raw ADU
